+++ packages/core/src/http/index.ts
@@ -1,5 +1,8 @@
+export * from './application/expressLikeApplication';
+export * from './handler/typedHandler';
 export * from './interfaces/expressLikeRouter.interface';
 export * from './openApiV3Generator/openApiV3Generator';
+export * from './router/expressLikeRouter';
 export * from './types/apiDefinition.types';
 export * from './types/contractDetails.types';
 export * from './types/router.types';
