+++ packages/express/tests/zod.forklaunch.express.test.ts
@@ -1,8 +1,8 @@
-import { ZodSchemaValidator, string } from '@forklaunch/validator/zod';
+import { SchemaValidator, string } from '@forklaunch/validator/zod';
 import { Server } from 'http';
 import { forklaunchExpress, forklaunchRouter } from '../forklaunch.express';
 
-const zodSchemaValidator = new ZodSchemaValidator();
+const zodSchemaValidator = SchemaValidator();
 
 describe('Forklaunch Express Tests', () => {
   let forklaunchApplication;
